Fixed bug in x86-64 when relocating segment headers on text section
One of the methods used to relocate the segment headers when there
is not room to expand them in-place is to relocate them at the
start of the .text section.  However, if the .text segment of a
binary is shorter than a full page, the rewriter would save a
backup of the .text section, followed by blank zeroes, and then
when restoring the .text section, it would overwrite the data that
came after the .text section with those zero bytes.  Since I check
whether the .text section is large enough to hold the segment
headers, I simply needed to stop restoring bytes after copying over
the number of bytes present in the .text section instead of just
copying a page's worth of bytes.  This is now fixed.